Worked on the CalculateWinningAccuracy tests

+++ MontyHallProblem/Helpers.cs
@@ -5,12 +5,12 @@ namespace MontyHallProblem
     public class Helpers
     {
 
-        public static decimal CalculateWinningAccuracy(int rightDecision, int noOfSimulations)
+        public static double CalculateWinningAccuracy(int rightDecision, int noOfSimulations)
         {
             if (noOfSimulations <= 0)
                 throw new Exception("You can't divide something with zero or lover");
 
-            decimal accuracyInProcent = (rightDecision / (decimal)noOfSimulations) * 100;
+            double accuracyInProcent = (rightDecision / (double)noOfSimulations) * 100;
 
             return Math.Round(accuracyInProcent);
         }
